Use macro for query initialization
Required for the traversalAsString value to be set automatically.

I've tried moving the queryInit().asInstanceOf[Query] to the method of a
separate object so that named parameters can be used and the casting at
the end is not needed, unfortunately the current macro takes the
argument's name as string instead of its expanded form. It might be
possible to create a new macro which doesn't do that and create a
prettier instantiation method, in the meantime this will have to do.